+++ packages/provider-sdk/src/errors/index.ts
@@ -1,5 +1,6 @@
 export { ProviderAbortError } from "./ProviderAbortError.js";
 export { ProviderError } from "./ProviderError.js";
 export { ProviderRequestError } from "./ProviderRequestError.js";
+export { ProviderResponseError } from "./ProviderResponseError.js";
 export { ProviderTimeoutError } from "./ProviderTimeoutError.js";
 export { ProviderUnavailableError } from "./ProviderUnavailableError.js";